Fix: 기본 답변 생성 로직 추가
