+++ src/components/Conversation/Response/ResponseChatChunk.tsx
@@ -6,7 +6,6 @@ import {
 } from "../../../Interface/Message/ResponseMessageType";
 import CardResponseMessage from "../../Messages/CardResponseMessage";
 import ContentResponseMessage from "../../Messages/ContentResponseMessage";
-import { DefaultCardResponseMessage } from "./DefaultCardResponseMessage";
 import { ResponseChat } from "./ResponseChat";
 
 export function ResponseChatChunk({
@@ -19,30 +18,26 @@ export function ResponseChatChunk({
 	const receivedResponse = message.content;
 	return (
 		<ResponseChat>
-			{receivedResponse ? (
-				receivedResponse.map(
-					(content: BasicResponseMessageType, idx: number) => {
-						switch (content.contentType) {
-							case "PlainText":
-								return (
-									<ContentResponseMessage
-										key={content.contentType + idx}
-										message={content as ContentResponseMessageType}
-									/>
-								);
-							case "ImageResponseCard":
-								return (
-									<CardResponseMessage
-										key={content.contentType + idx}
-										data={content as ImageResponseCardType}
-										messageID={`${messageID}-${idx}`}
-									/>
-								);
-						}
+			{receivedResponse.map(
+				(content: BasicResponseMessageType, idx: number) => {
+					switch (content.contentType) {
+						case "PlainText":
+							return (
+								<ContentResponseMessage
+									key={content.contentType + idx}
+									message={content as ContentResponseMessageType}
+								/>
+							);
+						case "ImageResponseCard":
+							return (
+								<CardResponseMessage
+									key={content.contentType + idx}
+									data={content as ImageResponseCardType}
+									messageID={`${messageID}-${idx}`}
+								/>
+							);
 					}
-				)
-			) : (
-				<DefaultCardResponseMessage messageID={messageID} />
+				}
 			)}
 		</ResponseChat>
 	);
